update: コンポーネント名を整理する
