+++ app/[lang]/generation/_components/editor-config-view/generation-config-sampler.tsx
@@ -19,7 +19,7 @@ type Props = {
   onChange(value: string): void
 }
 
-export const GenerationEditorConfigSampler = (props: Props) => {
+export const GenerationConfigSampler = (props: Props) => {
   return (
     <div className="flex flex-col gap-y-2">
       <div className="flex gap-x-2">
