+++ app/[lang]/generation/_components/editor-config-view/generation-config-scale.tsx
@@ -12,22 +12,20 @@ type Props = {
   onChange(value: number): void
 }
 
-/**
- * ClipSkipの設定
- * @returns
- */
-export const GenerationEditorConfigClipSkip = (props: Props) => {
+export const GenerationConfigScale = (props: Props) => {
   return (
     <div className="flex flex-col gap-y-2">
-      <div className="flex gap-x-2">
-        <span className="font-bold text-sm">{"ClipSkip"}</span>
+      <div className="flex gap-x-2 items-center">
+        <span className="font-bold text-sm">{"Scale"}</span>
         <TooltipProvider>
           <Tooltip>
             <TooltipTrigger asChild>
               <HelpCircleIcon className="w-4" />
             </TooltipTrigger>
             <TooltipContent>
-              <p>{"絵柄の雰囲気が変わります"}</p>
+              {
+                "Scale値が小さいほど創造的な画像を生成できます。値が大きいほど、より厳密にテキストを解釈します。"
+              }
             </TooltipContent>
           </Tooltip>
         </TooltipProvider>
@@ -36,7 +34,7 @@ export const GenerationEditorConfigClipSkip = (props: Props) => {
         type="number"
         value={props.value}
         min={1}
-        max={12}
+        max={15}
         onChange={(event) => {
           props.onChange(Number(event.target.value))
         }}
