+++ app/[lang]/generation/_components/editor-config-view/generation-config-step.tsx
@@ -12,7 +12,7 @@ type Props = {
   onChange(value: number): void
 }
 
-export const GenerationEditorConfigStep = (props: Props) => {
+export const GenerationConfigStep = (props: Props) => {
   return (
     <div className="flex flex-col gap-y-2">
       <div className="flex gap-x-2">
