--- app/[lang]/generation/_components/editor-config-view/generation-editor-config-open-favorite-model-toggle.tsx
@@ -1,21 +1,0 @@
-"use client"
-
-import { Toggle } from "@/components/ui/toggle"
-import { StarIcon } from "@radix-ui/react-icons"
-
-type Props = {
-  isActive: boolean
-}
-
-/**
- * お気に入りモデル一覧表示切替トグル
- * @param props
- * @returns
- */
-export const GenerationEditorConfigView = (props: Props) => {
-  return (
-    <Toggle>
-      <StarIcon className="w-4" />
-    </Toggle>
-  )
-}
