+++ app/[lang]/generation/_components/editor-submission-view/generation-submit-view.tsx
