+++ app/[lang]/generation/_components/editor-task-view/generation-image-download-button.tsx
