+++ app/[lang]/generation/_components/editor-task-view/generation-image-upload-button.tsx
