+++ app/[lang]/generation/_components/editor-task-view/generation-task-delete-button.tsx
