+++ app/[lang]/generation/_components/editor-task-view/generation-task-list-actions.tsx
@@ -1,11 +1,11 @@
 "use client"
 
-import { GenerationImageDownloadButton } from "@/app/[lang]/generation/_components/editor-task-view-view/generation-image-download-button"
-import { GenerationImagePostButton } from "@/app/[lang]/generation/_components/editor-task-view-view/generation-image-upload-button"
-import { GenerationTaskActionDropdownMenu } from "@/app/[lang]/generation/_components/editor-task-view-view/generation-task-action-dropdown-menu"
-import { GenerationTaskCountSelect } from "@/app/[lang]/generation/_components/editor-task-view-view/generation-task-count-select"
-import { GenerationTaskDeleteButton } from "@/app/[lang]/generation/_components/editor-task-view-view/generation-task-delete-button"
-import { GenerationTaskRatingSelect } from "@/app/[lang]/generation/_components/editor-task-view-view/generation-task-rating-select"
+import { GenerationImageDownloadButton } from "@/app/[lang]/generation/_components/editor-task-view/generation-image-download-button"
+import { GenerationImagePostButton } from "@/app/[lang]/generation/_components/editor-task-view/generation-image-upload-button"
+import { GenerationTaskActionDropdownMenu } from "@/app/[lang]/generation/_components/editor-task-view/generation-task-action-dropdown-menu"
+import { GenerationTaskCountSelect } from "@/app/[lang]/generation/_components/editor-task-view/generation-task-count-select"
+import { GenerationTaskDeleteButton } from "@/app/[lang]/generation/_components/editor-task-view/generation-task-delete-button"
+import { GenerationTaskRatingSelect } from "@/app/[lang]/generation/_components/editor-task-view/generation-task-rating-select"
 import { Button } from "@/components/ui/button"
 import { Toggle } from "@/components/ui/toggle"
 import { config } from "@/config"
