+++ app/[lang]/generation/_components/editor-task-view/generation-task-list-area.tsx
@@ -1,20 +1,16 @@
 "use client"
 
-import { GenerationEditorTaskList } from "@/app/[lang]/generation/_components/editor-task-view-view/generation-editor-task-list"
+import { GenerationTaskList } from "@/app/[lang]/generation/_components/editor-task-view/generation-task-list"
 import { useGenerationContext } from "@/app/[lang]/generation/_hooks/use-generation-context"
-import { InProgressGenerationCard } from "@/app/[lang]/generation/tasks/_components/in-progress-generation-card"
-import { ResponsivePagination } from "@/app/_components/responsive-pagination"
 import { useFocusTimeout } from "@/app/_hooks/use-focus-timeout"
 import { ScrollArea } from "@/components/ui/scroll-area"
 import { config } from "@/config"
 import { viewerImageGenerationTasksQuery } from "@/graphql/queries/viewer/viewer-image-generation-tasks"
 import { useQuery } from "@apollo/client"
-import { useState } from "react"
 import { toast } from "sonner"
 
 type Props = {
   sizeType?: string
-  isCreatingTasks: boolean
   rating: number
   isEditMode: boolean
   selectedTaskIds: string[]
@@ -23,6 +19,7 @@ type Props = {
   pcViewType?: string
   viewCount?: number
   setSelectedTaskIds: (selectedTaskIds: string[]) => void
+  onCancel?(): void
 }
 
 /**
@@ -30,26 +27,24 @@ type Props = {
  * @param props
  * @returns
  */
-export const GenerationTaskListHistory = (props: Props) => {
+export const GenerationTaskListArea = (props: Props) => {
   const context = useGenerationContext()
-
-  const [currentPage, setCurrentPage] = useState(1)
 
   const isTimeout = useFocusTimeout()
 
   const { data: tasks } = useQuery(viewerImageGenerationTasksQuery, {
     variables: {
-      limit: props.viewCount ?? 64,
-      offset: (currentPage - 1) * (props.viewCount ?? 0),
-      where: { minRating: 0 },
+      limit: 64,
+      offset: 0,
+      where: {},
     },
-    pollInterval: isTimeout ? undefined : 2000,
+    pollInterval: isTimeout ? 8000 : 2000,
   })
 
   const { data: ratingTasks } = useQuery(viewerImageGenerationTasksQuery, {
     variables: {
       limit: config.query.maxLimit,
-      offset: (currentPage - 1) * (props.viewCount ?? 0),
+      offset: 0,
       where: { minRating: 1 },
     },
   })
@@ -61,8 +56,6 @@ export const GenerationTaskListHistory = (props: Props) => {
   }
 
   const imageGenerationTasks = tasks.viewer?.imageGenerationTasks ?? []
-  const imageGenerationRatingTasks =
-    ratingTasks.viewer?.imageGenerationTasks ?? []
 
   /**
    * 非表示指定のタスクを除外
@@ -74,7 +67,7 @@ export const GenerationTaskListHistory = (props: Props) => {
   /**
    * フィルターしたレーティング済みタスク
    */
-  const currentRatingTasks = imageGenerationRatingTasks.filter((task) => {
+  const currentRatingTasks = imageGenerationTasks.filter((task) => {
     return (
       task.rating === props.rating &&
       task.nanoid &&
@@ -105,13 +98,17 @@ export const GenerationTaskListHistory = (props: Props) => {
 
   const activeTasks = currentTasks.filter((task) => {
     if (task.isDeleted || (!task.token && task.status === "DONE")) return false
-    return task.status === "IN_PROGRESS" || task.status === "DONE"
+    return (
+      task.status === "PENDING" ||
+      task.status === "IN_PROGRESS" ||
+      task.status === "DONE"
+    )
   })
 
   const activeRatingTasks = currentRatingTasks.filter((task) => {
-    if (!task || task.isDeleted || (!task.token && task.status === "DONE"))
-      return false
-    return task.status === "IN_PROGRESS" || task.status === "DONE"
+    if (task.isDeleted || (!task.token && task.status === "DONE")) return false
+    // return task.status === "IN_PROGRESS" || task.status === "DONE"
+    return task.status === "DONE"
   })
 
   const onSelectTask = (taskId: string | null, status: string) => {
@@ -143,55 +140,41 @@ export const GenerationTaskListHistory = (props: Props) => {
     if (props.sizeType === "full") {
       switch (size) {
         case "small":
-          return "p-2 grid grid-cols-3 gap-2 p-4 sm:pl-4 md:grid-cols-7 2xl:grid-cols-12 lg:grid-cols-10 xl:grid-cols-11"
+          return "p-2 grid grid-cols-3 gap-2 p-4 pt-0 sm:pl-4 md:grid-cols-7 2xl:grid-cols-12 lg:grid-cols-10 xl:grid-cols-11"
         case "middle":
-          return "p-2 grid grid-cols-2 gap-2 p-4 sm:pl-4 md:grid-cols-6 2xl:grid-cols-10 lg:grid-cols-8 xl:grid-cols-9"
+          return "p-2 grid grid-cols-2 gap-2 p-4 pt-0 sm:pl-4 md:grid-cols-6 2xl:grid-cols-10 lg:grid-cols-8 xl:grid-cols-9"
         case "big":
-          return "p-2 grid grid-cols-1 gap-2 p-4 sm:pl-4 md:grid-cols-4 2xl:grid-cols-5 lg:grid-cols-5 xl:grid-cols-5"
+          return "p-2 grid grid-cols-1 gap-2 p-4 pt-0 sm:pl-4 md:grid-cols-4 2xl:grid-cols-5 lg:grid-cols-5 xl:grid-cols-5"
         default:
-          return "p-2 grid grid-cols-2 gap-2 p-4 sm:pl-4 md:grid-cols-2 2xl:grid-cols-8 lg:grid-cols-5 xl:grid-cols-6"
+          return "p-2 grid grid-cols-2 gap-2 p-4 pt-0 sm:pl-4 md:grid-cols-2 2xl:grid-cols-8 lg:grid-cols-5 xl:grid-cols-6"
       }
     }
     switch (size) {
       case "small":
-        return "p-2 grid grid-cols-3 gap-2 p-4 sm:pl-4 md:grid-cols-3 2xl:grid-cols-5 lg:grid-cols-4 xl:grid-cols-3"
+        return "p-2 grid grid-cols-3 gap-2 p-4 pt-0 sm:pl-4 md:grid-cols-3 2xl:grid-cols-5 lg:grid-cols-4 xl:grid-cols-3"
       case "middle":
-        return "p-2 grid grid-cols-2 gap-2 p-4 sm:pl-4 md:grid-cols-2 2xl:grid-cols-4 lg:grid-cols-3 xl:grid-cols-2"
+        return "p-2 grid grid-cols-2 gap-2 p-4 pt-0 sm:pl-4 md:grid-cols-2 2xl:grid-cols-4 lg:grid-cols-3 xl:grid-cols-2"
       case "big":
-        return "p-2 grid grid-cols-1 gap-2 p-4 sm:pl-4 md:grid-cols-1 2xl:grid-cols-2 lg:grid-cols-1 xl:grid-cols-1"
+        return "p-2 grid grid-cols-1 gap-2 p-4 pt-0 sm:pl-4 md:grid-cols-1 2xl:grid-cols-2 lg:grid-cols-1 xl:grid-cols-1"
       default:
-        return "p-2 grid grid-cols-2 gap-2 p-4 sm:pl-4 md:grid-cols-2 2xl:grid-cols-4 lg:grid-cols-3 xl:grid-cols-2"
+        return "p-2 grid grid-cols-2 gap-2 p-4 pt-0 sm:pl-4 md:grid-cols-2 2xl:grid-cols-4 lg:grid-cols-3 xl:grid-cols-2"
     }
   }
 
   return (
-    <>
-      <ScrollArea>
-        <div className={getGridClasses(props.thumbnailSize)}>
-          {props.isCreatingTasks && (
-            <InProgressGenerationCard isCreatingTasks={true} />
-          )}
-          <GenerationEditorTaskList
-            tasks={props.rating === -1 ? activeTasks : activeRatingTasks}
-            isEditMode={props.isEditMode}
-            selectedTaskIds={props.selectedTaskIds}
-            pcViewType={pcViewType}
-            sizeType={props.thumbnailSize ?? "small"}
-            onRestore={onRestore}
-            onSelectTask={onSelectTask}
-          />
-        </div>
-      </ScrollArea>
-      {props.viewCount &&
-        tasks.viewer &&
-        tasks.viewer.remainingImageGenerationTasksTotalCount && (
-          <ResponsivePagination
-            perPage={props.viewCount}
-            maxCount={tasks.viewer.remainingImageGenerationTasksTotalCount}
-            currentPage={currentPage}
-            onPageChange={setCurrentPage}
-          />
-        )}
-    </>
+    <ScrollArea className="pb-64 md:pb-0">
+      <div className={`${getGridClasses(props.thumbnailSize)}`}>
+        <GenerationTaskList
+          tasks={props.rating === -1 ? activeTasks : activeRatingTasks}
+          isEditMode={props.isEditMode}
+          selectedTaskIds={props.selectedTaskIds}
+          pcViewType={pcViewType}
+          sizeType={props.thumbnailSize ?? "small"}
+          onRestore={onRestore}
+          onSelectTask={onSelectTask}
+          onCancel={props.onCancel}
+        />
+      </div>
+    </ScrollArea>
   )
 }
