+++ app/[lang]/generation/_components/editor-task-view/star-rating.tsx
